Windows support WIP:  - Initial support for Windows is in via OpenGL.  Not thoroughly tested,    and in fact it seems like there are a couple of discrepancies with the    OSX build (in particular the cone light object doesn't seem to be    rotated and I'm not sure it's operating at all).
   Removed the layout (std140) from the light struct definitions.

   Added WindowsGL.h/cpp which wraps getting gl functions for windows.  We
   use GenerateWindowsGL.pl to generate that source file, which uses as
   its data the output from a failing windows build.  That is to say, it's
   not that robust.

   Did a mess of double->float conversions to silence some warnings in
   RMath.  There are still a lot and some are worth fixing up.

   Had to add an option to shaders to allow the extensions, which required
   using a custom preprocessor with -o option to ignore unknown
   directives.  I'm not happy with the warning, may fork mcpp just so I
   can suppress that.  Or maybe write a simple replacement.

+++ Engine/Scene.cpp
@@ -508,5 +508,5 @@ int SceneGetSceneObjectsByType(SceneObject** dest, int size, Scene* scene, Scene
             *write++ = sceneObject;
     }
 
-    return write - first;
+    return (int) (write - first);
 }
